[#529] move from com.google.code.javaparser:javaparser:1.0.11 to com.github.javaparser:javaparser-core:2.5.1
