+++ jaxb-annox-parent/jaxb-annox/pom.xml
@@ -54,8 +54,8 @@
       <artifactId>commons-io</artifactId>
     </dependency>
     <dependency>
-      <groupId>com.google.code.javaparser</groupId>
-      <artifactId>javaparser</artifactId>
+      <groupId>com.github.javaparser</groupId>
+      <artifactId>javaparser-core</artifactId>
     </dependency>
     <dependency>
       <groupId>org.hibernate</groupId>
