+++ jaxb-annox-parent/jaxb-annox/src/main/java/org/jvnet/jaxb/annox/javaparser/AnnotationExprParser.java
@@ -1,11 +1,11 @@
-package org.jvnet.jaxb.annox.japa.parser;
+package org.jvnet.jaxb.annox.javaparser;
 
-import japa.parser.JavaParser;
-import japa.parser.ParseException;
-import japa.parser.ast.CompilationUnit;
-import japa.parser.ast.body.ClassOrInterfaceDeclaration;
-import japa.parser.ast.body.TypeDeclaration;
-import japa.parser.ast.expr.AnnotationExpr;
+import com.github.javaparser.JavaParser;
+import com.github.javaparser.ParseException;
+import com.github.javaparser.ast.CompilationUnit;
+import com.github.javaparser.ast.body.ClassOrInterfaceDeclaration;
+import com.github.javaparser.ast.body.TypeDeclaration;
+import com.github.javaparser.ast.expr.AnnotationExpr;
 
 import java.io.StringReader;
 import java.text.MessageFormat;
