+++ jaxb-annox-parent/jaxb-annox/src/main/java/org/jvnet/jaxb/annox/parser/XGenericFieldParser.java
@@ -1,6 +1,6 @@
 package org.jvnet.jaxb.annox.parser;
 
-import japa.parser.ast.expr.Expression;
+import com.github.javaparser.ast.expr.Expression;
 
 import java.lang.annotation.Annotation;
