+++ jaxb-annox-parent/jaxb-annox/src/main/java/org/jvnet/jaxb/annox/parser/XXAnnotationArrayAnnotationFieldParser.java
@@ -1,16 +1,16 @@
 //package org.jvnet.jaxb.annox.parser;
 //
-//import japa.parser.ast.Node;
-//import japa.parser.ast.expr.AnnotationExpr;
-//import japa.parser.ast.expr.ArrayInitializerExpr;
-//import japa.parser.ast.expr.Expression;
+//import com.github.javaparser.ast.Node;
+//import com.github.javaparser.ast.expr.AnnotationExpr;
+//import com.github.javaparser.ast.expr.ArrayInitializerExpr;
+//import com.github.javaparser.ast.expr.Expression;
 //
 //import java.lang.annotation.Annotation;
 //import java.text.MessageFormat;
 //import java.util.List;
 //
 //import org.jvnet.jaxb.annox.annotation.NoSuchAnnotationFieldException;
-//import org.jvnet.jaxb.annox.japa.parser.ast.visitor.AbstractGenericExpressionVisitor;
+//import visitor.ast.org.jvnet.jaxb.annox.javaparser.AbstractGenericExpressionVisitor;
 //import org.jvnet.jaxb.annox.model.XAnnotation;
 //import org.jvnet.jaxb.annox.model.annotation.field.XAnnotationField;
 //import org.jvnet.jaxb.annox.model.annotation.field.XArrayAnnotationField;
