+++ jaxb-annox-parent/jaxb-annox/src/main/java/org/jvnet/jaxb/annox/parser/java/visitor/BooleanExpressionVisitor.java
@@ -3,7 +3,7 @@ package org.jvnet.jaxb.annox.parser.java.visitor;
 import org.jvnet.jaxb.annox.model.annotation.value.XAnnotationValue;
 import org.jvnet.jaxb.annox.model.annotation.value.XBooleanAnnotationValue;
 
-import japa.parser.ast.expr.BooleanLiteralExpr;
+import com.github.javaparser.ast.expr.BooleanLiteralExpr;
 
 public final class BooleanExpressionVisitor extends
 		ExpressionVisitor<XAnnotationValue<Boolean>> {
