+++ jaxb-annox-parent/jaxb-annox/src/main/java/org/jvnet/jaxb/annox/parser/java/visitor/StringExpressionVisitor.java
@@ -3,7 +3,7 @@ package org.jvnet.jaxb.annox.parser.java.visitor;
 import org.jvnet.jaxb.annox.model.annotation.value.XAnnotationValue;
 import org.jvnet.jaxb.annox.model.annotation.value.XStringAnnotationValue;
 
-import japa.parser.ast.expr.StringLiteralExpr;
+import com.github.javaparser.ast.expr.StringLiteralExpr;
 
 public final class StringExpressionVisitor extends
 		ExpressionVisitor<XAnnotationValue<String>> {
