+++ jaxb-plugins-parent/jaxb-plugins/pom.xml
@@ -28,8 +28,8 @@
       <scope>provided</scope>
     </dependency>
     <dependency>
-      <groupId>com.google.code.javaparser</groupId>
-      <artifactId>javaparser</artifactId>
+      <groupId>com.github.javaparser</groupId>
+      <artifactId>javaparser-core</artifactId>
     </dependency>
     <dependency>
       <groupId>org.jvnet.jaxb</groupId>
