+++ pom.xml
@@ -420,9 +420,9 @@
 
       <!-- Java Parser -->
       <dependency>
-        <groupId>com.google.code.javaparser</groupId>
-        <artifactId>javaparser</artifactId>
-        <version>1.0.11</version>
+        <groupId>com.github.javaparser</groupId>
+        <artifactId>javaparser-core</artifactId>
+        <version>2.5.1</version>
       </dependency>
 
       <!-- JAXB Maven Plugin -->
